+++ src/vector.c
@@ -5,8 +5,8 @@ brh_vector3 vec3_rotate_x(brh_vector3 v, double angle)
 {
 	brh_vector3 rotated_vector = {
 		.x = v.x,
-		.y = v.y * cos(angle) - v.z * sin(angle),
-		.z = v.y * sin(angle) + v.z * cos(angle),
+		.y = v.y * cosf(angle) - v.z * sinf(angle),
+		.z = v.y * sinf(angle) + v.z * cosf(angle),
 	};
 	return rotated_vector;
 }
@@ -14,9 +14,9 @@ brh_vector3 vec3_rotate_x(brh_vector3 v, double angle)
 brh_vector3 vec3_rotate_y(brh_vector3 v, double angle)
 {
 	brh_vector3 rotated_vector = {
-		.x = v.x * cos(angle) + v.z * sin(angle),
+		.x = v.x * cosf(angle) + v.z * sinf(angle),
 		.y = v.y,
-		.z = -v.x * sin(angle) + v.z * cos(angle),
+		.z = -v.x * sinf(angle) + v.z * cosf(angle),
 	};
 
 	return rotated_vector;
@@ -25,8 +25,8 @@ brh_vector3 vec3_rotate_y(brh_vector3 v, double angle)
 brh_vector3 vec3_rotate_z(brh_vector3 v, double angle)
 {
 	brh_vector3 rotated_vector = {
-		.x = v.x * cos(angle) - v.y * sin(angle),
-		.y = v.x * sin(angle) + v.y * cos(angle),
+		.x = v.x * cosf(angle) - v.y * sinf(angle),
+		.y = v.x * sinf(angle) + v.y * cosf(angle),
 		.z = v.z,
 	};
